+++ src/units/Angle.cpp
@@ -10,11 +10,11 @@ Angle::Angle() : m_radians(0) {
 Angle::~Angle() {
 }
 
-float Angle::getRadians() const {
+double Angle::getRadians() const {
     return fmod(m_radians + 2*M_PI, 2*M_PI); // Ensure [0, 2pi)
 }
 
-float Angle::getDegrees() const {
+double Angle::getDegrees() const {
     return getRadians() * 360 / (2 * M_PI);
 }
