+++ src/units/Degrees.h
@@ -8,11 +8,11 @@ class Degrees : public Angle {
 
 public:
     Degrees();
-    Degrees(float degrees);
+    Degrees(double degrees);
     Degrees(const Angle& angle);
     Degrees operator+(const Angle& angle) const;
     Degrees operator-(const Angle& angle) const;
-    Degrees operator*(float factor) const;
+    Degrees operator*(double factor) const;
 
 };
