+++ src/units/Meters.cpp
@@ -6,7 +6,7 @@ Meters::Meters() {
     m_meters = 0.0;
 }
 
-Meters::Meters(float meters) {
+Meters::Meters(double meters) {
     m_meters = meters;
 }
 
@@ -22,15 +22,15 @@ Meters Meters::operator-(const Distance& distance) const {
     return Meters(getMeters() - distance.getMeters());
 }
 
-Meters Meters::operator*(float factor) const {
+Meters Meters::operator*(double factor) const {
     return Meters(getMeters() * factor);
 }
 
-Meters Meters::operator/(float factor) const {
+Meters Meters::operator/(double factor) const {
     return Meters(getMeters() / factor);
 }
 
-float Meters::operator/(const Distance& distance) const {
+double Meters::operator/(const Distance& distance) const {
     return getMeters() / distance.getMeters();
 }
