+++ src/units/MetersPerSecond.h
@@ -7,7 +7,7 @@ namespace sim {
 class MetersPerSecond : public Speed {
 
 public:
-    MetersPerSecond(float metersPerSecond);
+    MetersPerSecond(double metersPerSecond);
     MetersPerSecond(const Speed& speed);
     MetersPerSecond operator+(const Speed& speed) const;
     MetersPerSecond operator-(const Speed& speed) const;
